option changes

+++ programs/e2spt_evalrefine_gui.py
@@ -22,6 +22,7 @@ def main():
 	parser.add_argument("--jsout", type=str,help="json output", default=None)
 	parser.add_argument("--inplace", action="store_true", default=False ,help="overwrite input.")
 	parser.add_argument("--invert", action="store_true", default=False ,help="invert direction.")
+	parser.add_argument("--readonly", action="store_true", default=False ,help="read only mode.")
 	parser.add_argument("--mode", type=str,help="choose from rad/line, default is line", default="line")
 	parser.add_argument("--vec", type=str,help="vector direction, default 0,0,1", default="0,0,1")
 
@@ -70,47 +71,66 @@ class EMSptEval(QtWidgets.QMainWindow):
 		self.gbl.addWidget(self.imglst, 0,1,10,1)
 		self.imglst.cellClicked[int, int].connect(self.on_list_selected)
 		
-		self.bt_save=QtWidgets.QPushButton("Save")
-		self.gbl.addWidget(self.bt_save, 0,0,1,1)
-		self.bt_save.clicked[bool].connect(self.save_json)
 
 		self.update_list()
 		self.plotwiny=MplCanvas(self)
 		self.plotwiny.setMinimumSize(500, 500)
 		self.ploty=self.plotwiny.axes
 		self.gbl.addWidget(self.plotwiny, 0,2,7,1)
-		self.plotwiny.mpl_connect('button_press_event', self.onclick_ploty)
+		
 		
 		self.plotwinz=MplCanvas(self)
 		self.plotwiny.setMinimumSize(500, 200)
 		self.plotz=self.plotwinz.axes
 		self.gbl.addWidget(self.plotwinz, 7,2,3,1)
-		self.plotwinz.mpl_connect('button_press_event', self.onclick_plotz)
 		self.cursel=-1
 		
+		if not options.readonly:
+			self.bt_save=QtWidgets.QPushButton("Save")
+			self.gbl.addWidget(self.bt_save, 0,0,1,1)
+			self.bt_save.clicked[bool].connect(self.save_json)
+			self.plotwiny.mpl_connect('button_press_event', self.onclick_ploty)
+			self.plotwinz.mpl_connect('button_press_event', self.onclick_plotz)
 		
 	def load_json(self):
-		
-		self.jsdict=js=dict(js_open_dict(self.jsname)).copy()
-		keys=js.keys()
-		self.data=[]
-		print("Reading particles from json...")
-		for k in keys:
-			src, ii = eval(k)
-			e=EMData(src, ii, True)
-			srcname=e["file_twod"]#e["data_source"]
-			self.data.append({"k":k,
-					"srcfull":srcname,
-					"src":base_name(srcname),
-					"srci":e["data_n"],
-					"pos":e["ptcl_source_coord"],
-					"xf":js[k]["xform.align3d"], 
-					"score":js[k]["score"],
-					"toinv":False,
-					})
-			sys.stdout.write("\r  {}/{}        ".format(len(self.data), len(keys)))
-			sys.stdout.flush()
-		print()
+		if self.jsname.endswith('.json'):
+			self.jsdict=js=dict(js_open_dict(self.jsname)).copy()
+			keys=js.keys()
+			self.data=[]
+			print("Reading particles from json...")
+			for k in keys:
+				src, ii = eval(k)
+				e=EMData(src, ii, True)
+				srcname=e["file_twod"]#e["data_source"]
+				self.data.append({"k":k,
+						"srcfull":srcname,
+						"src":base_name(srcname),
+						"pos":e["ptcl_source_coord"],
+						"xf":js[k]["xform.align3d"], 
+						"toinv":False,
+						})
+				sys.stdout.write("\r  {}/{}        ".format(len(self.data), len(keys)))
+				sys.stdout.flush()
+			print()
+		else:
+			self.data=[]
+			print("Reading particles from lst...")
+			self.jsdict=lst=load_lst_params(self.jsname)
+			for k,ln in enumerate(lst):
+				e=EMData(ln["src"], ln["idx"], True)
+				srcname=e["file_twod"]#e["data_source"]
+				self.data.append({"k":k,
+						"srcfull":srcname,
+						"src":base_name(srcname),
+						"pos":e["ptcl_source_coord"],
+						"xf":ln["xform.align3d"], 
+						"toinv":False,
+						})
+				sys.stdout.write("\r  {}/{}        ".format(len(self.data), len(lst)))
+				sys.stdout.flush()
+			print()
+			
+			
 		
 		fs=natural_sort([d["src"] for d in self.data])
 		files, count=np.unique(fs, return_counts=True)
@@ -133,29 +153,45 @@ class EMSptEval(QtWidgets.QMainWindow):
 		elif self.options.jsout:
 			outname=self.options.jsout
 		else:
-			outname=self.jsname.replace(".json", "_mod.json")
+			if self.jsname.endswith('.json'):
+				outname=self.jsname.replace(".json", "_mod.json")
+			else:
+				outname=self.jsname.replace(".lst", "_mod.lst")
 			
 		print("Writting output to {}".format(outname))
 		rot=Transform({"type":"eman","alt":180, "az":0})
-		jsout={}
-		suminv=0
-		for d in self.data:
-			k=d["k"]
-			
-			val=self.jsdict[k].copy()
-			if d["toinv"]:
-				suminv+=1
-				xf=val["xform.align3d"]
-				val["xform.align3d"]=rot*xf
+		if self.jsname.endswith('.json'):
+			jsout={}
+			suminv=0
+			for d in self.data:
+				k=d["k"]
+				val=self.jsdict[k].copy()
+				if d["toinv"]:
+					suminv+=1
+					xf=val["xform.align3d"]
+					val["xform.align3d"]=rot*xf
+					
+				jsout[k]=val
 				
-			jsout[k]=val
+			if os.path.isfile(outname):
+				os.remove(outname)
+			js=js_open_dict(outname)
+			js.update(jsout)
+			js.close()
+			print("{} particles inverted".format(suminv))
 			
-		if os.path.isfile(outname):
-			os.remove(outname)
-		js=js_open_dict(outname)
-		js.update(jsout)
-		js.close()
-		print("{} particles inverted".format(suminv))
+		else:
+			lstout=[l.copy() for l in self.jsdict]
+			suminv=0
+			for k, d in enumerate(self.data):
+				if d["toinv"]:
+					suminv+=1
+					xf=lstout[k]["xform.align3d"]
+					lstout[k]["xform.align3d"]=rot*xf
+					
+			save_lst_params(lstout, outname)
+			print("{} particles inverted".format(suminv))
+			
 		
 		for fname in self.filenames:
 			v=self.fileinfo[fname]["point"]
@@ -181,8 +217,28 @@ class EMSptEval(QtWidgets.QMainWindow):
 		
 	def on_list_selected(self, row, col):
 		self.cursel=idx=int(self.imglst.item(row, 0).text())
-		self.check_orient(self.filenames[idx], True)
+		if self.options.readonly:
+			self.do_plot(self.filenames[idx])
+		else:
+			self.check_orient(self.filenames[idx], True)
 		
+	def do_plot(self, fname):
+		data=[d for d in self.data if d["src"]==fname]
+		
+		pos=np.array([d["pos"] for d in data])
+		ks=[d["k"] for d in data]
+		xfs=[Transform(d["xf"]) for d in data]
+		for x in xfs:
+			x.set_trans(0,0,0)
+			x.invert()
+		vecs=np.array([xf.transform(self.vec) for xf in xfs])
+		
+		for j,plot in enumerate([self.ploty, self.plotz]):
+			plot.cla()
+			plot.quiver(pos[:,0], pos[:,j+1], vecs[:,0], vecs[:,j+1], scale=50, color='b',width=.01)
+				
+		self.plotwiny.draw()
+		self.plotwinz.draw()
 		
 	def check_orient(self, fname, doplot=True):
 		data=[d for d in self.data if d["src"]==fname]
@@ -194,7 +250,6 @@ class EMSptEval(QtWidgets.QMainWindow):
 			x.set_trans(0,0,0)
 			x.invert()
 		vecs=np.array([xf.transform(self.vec) for xf in xfs])
-		#vecs=np.array([xf.transform([-.44,-.89,0]) for xf in xfs])
 		pc=np.mean(pos, axis=0)
 		
 		glbpt=self.fileinfo[fname]["point"]
